feat(dasclaw_core): move context::memory into dasclaw_core (ADR-152 §3 F3.6 slice 1)
ADR-152 §3 F3.6 slice 1 / 6. Moves desktop-client/ironclaw/src/context/memory.rs
verbatim into dasclaw_core::context::memory; this is the only file under the
context/ family that has zero same-level deps and only needs an import rebase
from `crate::llm::{ChatMessage,Role}` to `crate::messages::{ChatMessage,Role}`
(both types already live in dasclaw_core).

Changes
- git mv desktop-client/ironclaw/src/context/memory.rs -> crates/dasclaw_core/src/context/memory.rs
- crates/dasclaw_core/Cargo.toml: add rust_decimal = { version = "1", features = ["serde"] }
  (required by ActionRecord.cost / Memory.total_cost; no other api change)
- crates/dasclaw_core/src/lib.rs: register pub mod context
- crates/dasclaw_core/src/context/mod.rs: new file, re-exports memory::{ActionRecord, ConversationMemory, Memory}
- desktop-client/ironclaw/src/context/mod.rs: drop `mod memory;` and pull the
  three names back via `pub use dasclaw_core::context::memory::{...};`. All
  pre-existing `crate::context::Memory` call sites (worker::job, etc.) stay
  source-compatible through this re-export.
- desktop-client/ironclaw/src/context/fallback.rs: switch local imports from
  the now-absent `crate::context::memory::Memory` to the parent re-export
  `crate::context::Memory`.

Verbatim-port compliance (ADR-129 §1.3)
- Only edits inside the moved file are the four `crate::llm::*` ->
  `crate::messages::*` import rebases. No logic, no naming, no doc, no test
  rewrites.

Verification (local)
- cargo check -p dasclaw_core --tests        clean
- cargo check -p dasclaw --tests             clean
- cargo nextest run -p dasclaw_core <memory> 21/21 pass
- cargo fmt --all                            clean
- python3.12 scripts/check_no_panics.py      clean
- cargo clippy --no-deps -p dasclaw_core --all-targets -- -D warnings  clean

Next slice
- F3.6 slice 2 candidates: JobError extraction into dasclaw_core::error, then
  manager.rs (consumes Memory + JobError). state.rs / fallback.rs remain
  blocked on the JobContext god-struct split (ADR-152 §11.7 follow-up).
