+++ crates/dasclaw_core/Cargo.toml
@@ -26,6 +26,8 @@ tracing = "0.1"
 # session / undo / task 依赖
 chrono = { version = "0.4", features = ["serde"] }
 uuid = { version = "1", features = ["v4", "v5", "serde"] }
+# context::memory 依赖 (ADR-152 §3 F3.6 slice 1, ActionRecord.cost / Memory.total_cost)
+rust_decimal = { version = "1", features = ["serde"] }
 # session 的 truncate_preview helper
 ironclaw_common = { path = "../../desktop-client/ironclaw/crates/ironclaw_common", version = "0.1.0" }
 # session_manager 需要 RwLock/Mutex/spawn (fire-and-forget hooks)
